finalization2

+++ app/src/main/java/com/dicoding/asclepius/helper/ImageClassifierHelper.kt
@@ -57,7 +57,7 @@ class ImageClassifierHelper(
         val contentResolver = context.contentResolver
         val imageProcessor = ImageProcessor.Builder()
             .add(ResizeOp(224, 224, ResizeOp.ResizeMethod.NEAREST_NEIGHBOR))
-            .add(CastOp(DataType.UINT8))
+            .add(CastOp(DataType.FLOAT32))
             .build()
 
         val bitmap = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
